# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI QA Interview Coach API
  version: 0.1.0
  description: Initial MVP contract for interview practice, feedback, CRI and diary flows.
servers:
  - url: http://localhost:3001/api/v1
paths:
...
  /interviews:
    post:
      summary: Start interview session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartInterviewRequest"
      responses:
        "201":
          description: Interview session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{sessionId}/answers:
    post:
      summary: Submit interview answer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerRequest"
      responses:
        "201":
          description: Answer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{sessionId}/complete:
    post:
      summary: Complete interview session
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "201":
          description: Interview session completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    StartInterviewRequest:
      type: object
      required: [language, targetRole, seniority, topic, difficulty]
      properties:
        language:
          type: string
          enum: [pt-BR, en]
        targetRole:
          type: string
        seniority:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [medium, advanced, specialist]
        interviewerStyle:
          type: string
    InterviewSession:
      type: object
      properties:
        id:
          type: string
        language:
          type: string
        status:
          type: string
        currentQuestion:
          type: string
        turns:
          type: array
          items:
            $ref: "#/components/schemas/InterviewTurn"
    InterviewTurn:
      type: object
      properties:
        orderIndex:
          type: number
        question:
          type: string
        answer:
          type: string
        coachNote:
          type: string
    SubmitAnswerRequest:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
...